move ScriptFileChooser to /swing, by convention
