+++ java/src/jmri/jmrit/jython/RunJythonScript.java
@@ -6,7 +6,7 @@ import javax.script.ScriptException;
 import javax.swing.Icon;
 import javax.swing.JFileChooser;
 import jmri.script.JmriScriptEngineManager;
-import jmri.script.ScriptFileChooser;
+import jmri.script.swing.ScriptFileChooser;
 import jmri.util.FileUtil;
 import jmri.util.swing.JmriAbstractAction;
 import jmri.util.swing.WindowInterface;
