+++ java/src/jmri/script/swing/Bundle.properties
@@ -1,0 +1,1 @@
+
