+++ java/test/jmri/script/swing/ScriptFileChooserTest.java
@@ -1,4 +1,4 @@
-package jmri.script;
+package jmri.script.swing;
 
 import jmri.util.JUnitUtil;
